renamed graph-validation stuff with suffix 'runner(s)'; move helper functions to a utils module
